app/requirements.txt
pandas
autogluon.tabular
streamlit
plotly
